Rules text box measure
Added debug information for Body height measure. If this measured value (height of text after all automatic adjustments of line spacing and font scaling) is greater than the height defined in the region, then the red error box is drawn on the output image.

Note that currently you are drawing the trait line as paragraph, so you will see two reports of this measurements. Also, this won't work on card title because those are drawn as lines.

Also, I've added all the necessary files to make the plugin project directly "openable" in Strange Eons for easier coding, and updated the description to note that this is a custom version for Alep.